pyproject.toml
[tool.poetry]
name = "burpless"
version = "0.1.0"
description = "A gherkin parser that uses LALR instead of regex for parsing"
authors = ["nomorepanic <nomorepanic@strangemachines.io>"]
license = "GPL-3.0"
include = ["LICENSE", "README.md"]
repository = "https://github.com/strangemachines/burpless"
classifiers = [
    'Intended Audience :: Developers'
]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
lark-parser = "^0.7.2"


[tool.poetry.dev-dependencies]
flake8 = "^3.7"
flake8-blind-except = "~0.1.1"
flake8-broken-line = "~0.1.1"
flake8-builtins = "~1.4"
flake8-quotes = "~2.1"
flake8-import-order = "~0.18.1"
ipython = "~7.8"
pep8-naming = "~0.8.2"
pytest = "^3.0"
pytest-mock = "^1.10"
pytest-sugar = "^0.9.2"


[tool.poetry.scripts]
burpless = 'burpless.Cli:Cli.main'


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
